Problem: can't build linux package under macOS
The problem is generally solved by running inside the specialized `pgpm`
container which has necessary dependencies installed.

However, this requires us to run it there.

Solution: podmanize everything!

We already use podman to bring components we may not have – like rust
toolchain. Why not use it to bring `mock` in?

This is not perfect yet, and relies on volume mapping (which is what we do
already anyway) – for which we had to do special accommodations to ensure
we know these paths and they are absolute.

We currently use the `pgpm` image as the one that should contain `mock`. But
may be it should be a different, smaller image?

To build for a particular OS, use the `--os` flag, for example:

```
--os rocky+epel-9
```