Problem: building extensions that require other extensions
This doesn't currently work because we try to install those other extensions in
the last build and they are not available.

Solution: build them together with the package

The implementation is currently very, very hacky – that Proc refinement is just
a massive hack. I've also fixed rebuild chaining but it is extremely brittle as
well.

But at least it works. I think.