Problem: rpm throws an interesting exception
```
[ 12/135] Installing libpcap-14:1.10.5-1.fc41.aarch64 100% | 501.3 KiB
[ 13/135] Installing libfdisk-0:2.40.4-1.fc41.aarch64 100% | 413.4 KiBterminate called after throwing an instance of 'std::length_error'
  what():  basic_string::_M_replace_aux
```

This didn't happen before

Solution: don't allocate pseudo-terminal in podman

We don't really need it. Not clear why it was working before, though.

And not clear what kind of problem rpm runs into when it think it's an
interactive terminal, but it's a captured terminal.

But now it seems to work.