Problem: pbuilder uses tar, which results in "Cannot readlink: Invalid argument"
The issue is present on podman hosts when they are themslves KVM guests, but
is not present if podman is used on host OS, which iteself is installed on
bare metal.

Solution: create `faketar` script and patch pbuilder to use it instead of tar.
The `faketar` script uses simple `cp` and `mv` commands, which not only help
us avoid the issue with `tar`, but also make our builds faster by skipping
compression & decompression of tgz.

Due to a rather esoteric nature of the problem `faketar` script is reasonable
workaround, which also saves on build time.